Emote UI (#100)
features
- add props and sounds to emotes
- add emote selection ui page
- clean up default emotes loading (still without props/sounds)
- improve hide/force render avatar computations
- remove console, route via chat

misc
- rework collectible urn api
- rework loaders for emotes and wearables

bugfixes
- fix crash with no audio device - check for device before playing sounds
- freeze ffmpeg version to fix macos build

+++ src/main.rs
@@ -84,7 +84,18 @@ fn main() {
                 .join(f.path().file_stem().unwrap())
         });
 
-    File::create(format!("{}.touch", SESSION_LOG.get().unwrap())).unwrap();
+    let mut args = pico_args::Arguments::from_env();
+
+    let file_log = !args.contains("--console") && !cfg!(feature = "tracy");
+
+    if file_log {
+        File::create(SESSION_LOG.get().unwrap())
+            .expect("failed to create log file")
+            .write_all(format!("{}\n\n", SESSION_LOG.get().unwrap()).as_bytes())
+            .expect("failed to create log file");
+
+        File::create(format!("{}.touch", SESSION_LOG.get().unwrap())).unwrap();
+    }
 
     // warnings before log init must be stored and replayed later
     let mut warnings = Vec::default();
@@ -98,7 +109,6 @@ fn main() {
                 .ok()
         })
         .unwrap_or_default();
-    let mut args = pico_args::Arguments::from_env();
 
     let final_config = AppConfig {
         server: args
@@ -226,27 +236,24 @@ fn main() {
                 })
                 .set(bevy::log::LogPlugin {
                     filter: "wgpu=error,naga=error".to_string(),
-                    update_subscriber: Some(
-                        move |_subscriber: BoxedSubscriber| -> BoxedSubscriber {
-                            #[cfg(not(feature = "tracy"))]
-                            {
-                                let (non_blocking, guard) = tracing_appender::non_blocking(
-                                    File::options()
-                                        .write(true)
-                                        .open(SESSION_LOG.get().unwrap())
-                                        .unwrap(),
-                                );
-                                let l = bevy::log::tracing_subscriber::fmt()
-                                    .with_ansi(false)
-                                    .with_writer(non_blocking)
-                                    .finish();
-                                Box::leak(Box::new(guard));
-                                Box::new(l)
-                            }
-                            #[cfg(feature = "tracy")]
-                            _subscriber
-                        },
-                    ),
+                    update_subscriber: if file_log {
+                        Some(move |_subscriber: BoxedSubscriber| -> BoxedSubscriber {
+                            let (non_blocking, guard) = tracing_appender::non_blocking(
+                                File::options()
+                                    .write(true)
+                                    .open(SESSION_LOG.get().unwrap())
+                                    .unwrap(),
+                            );
+                            let l = bevy::log::tracing_subscriber::fmt()
+                                .with_ansi(false)
+                                .with_writer(non_blocking)
+                                .finish();
+                            Box::leak(Box::new(guard));
+                            Box::new(l)
+                        })
+                    } else {
+                        None
+                    },
                     ..default()
                 })
                 .build()
@@ -333,7 +340,9 @@ fn main() {
     log_panics::init();
     app.run();
 
-    std::fs::remove_file(format!("{}.touch", SESSION_LOG.get().unwrap())).unwrap();
+    if file_log {
+        std::fs::remove_file(format!("{}.touch", SESSION_LOG.get().unwrap())).unwrap();
+    }
 }
 
 fn setup(
